Fixing planning for order sensitive window functions
Order sensitive windows functions such as `last_value` can produce wrong
results if the plan does not respect the given sort order and if there
is motion in between of resorting.

If the input plan specifies a specific order `s1` and the window sorting
needs to apply sort order of `s2` then:

If `s2` is a prefix of `s1`, window sorting has to made on `s1` to
eleminate wrong resuls in order sensitive window functions.

Otherwise, window sorting can go with the normal flow and sort based on
`s2`.

Signed-off-by: Omer Arap <oarap@pivotal.io>