Enable full scan hyperloglog to calculate root NDV
This commit introduces a feature to analyze the table with a keyword
`fullscan` to generate hyperloglog counter by scanning the entire
table once. While this will slow down the analyze, it will produce
accurate estimate for the NDV in statistics. When we merge the leaf
table statistics for generating root level statistics, we are going to
have a NDV very close to the value we can get without rescanning entire
leaf partitions.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>