Handle merge stats with empty HLL stats
For varlena columns, if the data in the sample is greater than the
`WIDTH_THRESHOLD` value, analyze does not create HLL stats for the leaf of
those columns. This commit handles merging of leaf stats when the leaf
partitions are analyzed but there are no HLL stats available for
merging. In such cases, it merges root stats similar to the way it
computes leaf stats for such columns.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>