Fix gpsd to account newly added column to pg_statistic
Recent postgres merge introduced a new column `stainherit` to
`pg_statistic` table. When `gpsd` dumps the pg_statistic not only a
column is missing but also all the other columns are shifted and saved
in wrong column index. This causes failure on reloading the gpsd.

This commit fixes the issue and adds the missing column.