Option for analyze to collect minimal stats for temp tables
If a user has a case where they have large external tables and their
loading process uses temp tables. ANALYZE takes a long time to run on
those tables. If we do not ANALYZE them, ORCA has no idea about size of
the tables and produced slow plans. Therefore it is better to provide at
least the size of the table which requires ANALYZE to just collect
reltuples and relpages and not rest of the stats.

We provided a GUC `gp_statistics_reltuples_for_temp_tables_only` and
when turned on ANALYZE will only collect reltuples and relpages info and
not any other statistics for temp tables.

Signed-off-by: Omer Arap <oarap@pivotal.io>
Signed-off-by: Sambitesh Dash <sdash@pivotal.io>