Autotrigger stats merge for root partition when leaf is analyzed
This commit automatically triggers root partition statistics merge
operation when the leaf partition is analyzed individually. It only
triggers the merge only if the other partitions are already analyzed
and stats are in place.

Remove analyze MERGE

As we are automatically trigger root partition merge from theh previous
commit, we do not need the separate keyword to the root partition
statistics MERGE.