Fix auto merge root partition stats
This commit fixes couple of issues for root table statistics merge when
a single partition is analyzed.

1. Auto merge was not functional when a column list is specified in the
analyze command such as `ANALYZE part1 (a)`.

2. When there is misalignment for attnums between partitions, it was not
checking the availability of stats based on the attnum which is not
aligned and discovered properly.

3. Auto merge should check every column's stats availability for each
partition. `forboth` macro was used in a wrong way as if it was a double
for loop which was causing auto merge not to be function when the number
of partitions is not equal to the number of columns.

This commit fixes all these issues.