Fix ANALYZE ROOTPARTITION, always collect sample
This commit fixes the behavior of `ANALYZE ROOTPARTITION` where we
should always collect samples even when there are stats available in the
leaf partition for merging.

We also do significant refactoring and with this commit as we moved most
of the incremental analyze code to the `anaalyzeutils.c`.  # Please
enter the commit message for your changes. Lines starting